seattle/bnn/requirements.txt
pandas>=2.0
numpy==1.24.4
matplotlib>=3.6
scikit-learn>=1.2
seaborn>=0.12
pymc>=5.10
arviz>=0.16
pytensor>=2.16
jaxlib>=0.4.15+cuda12.cudnn89
